Feat: #63 MSW handler 구조 변경
