+++ src/mocks/server.ts
@@ -1,5 +1,5 @@
 import { setupServer } from 'msw/node';
-import { handlers } from './handlers';
+import handlers from '@mocks/handlers';
 
 export const server = setupServer(...handlers);
 server.events.on('request:start', ({ request }) => {
